Moved some types from types.h to osqp_api_types.h
